debug command facelift.
