+++ src/commands.rs
@@ -16,7 +16,6 @@ pub(crate) mod count;
 pub(crate) mod cp;
 pub(crate) mod date;
 pub(crate) mod debug;
-pub(crate) mod debug_value;
 pub(crate) mod echo;
 pub(crate) mod enter;
 pub(crate) mod env;
@@ -103,7 +102,7 @@ pub(crate) use config::Config;
 pub(crate) use count::Count;
 pub(crate) use cp::Cpy;
 pub(crate) use date::Date;
-pub(crate) use debug_value::DebugValue;
+pub(crate) use debug::Debug;
 pub(crate) use echo::Echo;
 pub(crate) use enter::Enter;
 pub(crate) use env::Env;
