+++ src/commands/debug.rs
@@ -1,8 +1,10 @@
 use crate::commands::WholeStreamCommand;
-use crate::errors::ShellError;
 use crate::prelude::*;
 
 pub struct Debug;
+
+#[derive(Deserialize)]
+pub struct DebugArgs {}
 
 impl WholeStreamCommand for Debug {
     fn name(&self) -> &str {
@@ -14,7 +16,7 @@ impl WholeStreamCommand for Debug {
     }
 
     fn usage(&self) -> &str {
-        "Debug input fed."
+        "Print the Rust debug representation of the values"
     }
 
     fn run(
@@ -22,18 +24,19 @@ impl WholeStreamCommand for Debug {
         args: CommandArgs,
         registry: &CommandRegistry,
     ) -> Result<OutputStream, ShellError> {
-        debug(args, registry)
+        args.process(registry, debug_value)?.run()
     }
 }
 
-pub fn debug(args: CommandArgs, _registry: &CommandRegistry) -> Result<OutputStream, ShellError> {
-    let input = args.input;
+fn debug_value(
+    _args: DebugArgs,
+    RunnableContext { mut input, .. }: RunnableContext,
+) -> Result<impl ToOutputStream, ShellError> {
+    let stream = async_stream! {
+        while let Some(row) = input.values.next().await {
+            yield ReturnSuccess::debug_value(row.clone())
+        }
+    };
 
-    Ok(input
-        .values
-        .map(|v| {
-            outln!("{:?}", v);
-            ReturnSuccess::value(v)
-        })
-        .to_output_stream())
+    Ok(stream)
 }
